backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
passlib[bcrypt]
python-dotenv
pydantic
email-validator
cryptography
python-multipart
psycopg2
psycopg2-binary